Fix paths associated with rename
