+++ apps/android_deploy/app/src/main/jni/tvm_runtime.h
@@ -32,7 +32,7 @@
 #include "../src/runtime/cpu_device_api.cc"
 #include "../src/runtime/dso_library.cc"
 #include "../src/runtime/file_utils.cc"
-#include "../src/runtime/graph/graph_runtime.cc"
+#include "../src/runtime/graph_executor/graph_runtime.cc"
 #include "../src/runtime/library_module.cc"
 #include "../src/runtime/module.cc"
 #include "../src/runtime/ndarray.cc"
