+++ apps/bundle_deploy/runtime.cc
@@ -26,7 +26,7 @@
 #include "../../src/runtime/container.cc"
 #include "../../src/runtime/cpu_device_api.cc"
 #include "../../src/runtime/file_utils.cc"
-#include "../../src/runtime/graph/graph_runtime.cc"
+#include "../../src/runtime/graph_executor/graph_runtime.cc"
 #include "../../src/runtime/library_module.cc"
 #include "../../src/runtime/module.cc"
 #include "../../src/runtime/ndarray.cc"
