+++ apps/ios_rpc/tvmrpc/TVMRuntime.mm
@@ -46,7 +46,7 @@
 #include "../../../src/runtime/rpc/rpc_session.cc"
 #include "../../../src/runtime/rpc/rpc_socket_impl.cc"
 // Graph runtime
-#include "../../../src/runtime/graph/graph_runtime.cc"
+#include "../../../src/runtime/graph_executor/graph_runtime.cc"
 // Metal
 #include "../../../src/runtime/metal/metal_device_api.mm"
 #include "../../../src/runtime/metal/metal_module.mm"
