+++ cmake/modules/CUDA.cmake
@@ -73,7 +73,7 @@ if(USE_CUDA)
       message(FATAL_ERROR "CUDA Graph requires CUDA 10 or above, got=" ${CUDAToolkit_VERSION})
     endif()
     message(STATUS "Build with Graph runtime with CUDA Graph support...")
-    file(GLOB RUNTIME_CUDA_GRAPH_SRCS src/runtime/graph/cuda_graph/*.cc)
+    file(GLOB RUNTIME_CUDA_GRAPH_SRCS src/runtime/graph_executor/cuda_graph/*.cc)
     list(APPEND RUNTIME_SRCS ${RUNTIME_CUDA_GRAPH_SRCS})
   endif()
 else(USE_CUDA)
