+++ golang/src/tvm_runtime_pack.cc
@@ -43,7 +43,7 @@
 #include "src/runtime/system_library.cc"
 
 // Graph runtime
-#include "src/runtime/graph/graph_runtime.cc"
+#include "src/runtime/graph_executor/graph_runtime.cc"
 
 // Uncomment the following lines to enable RPC
 // #include "../../src/runtime/rpc/rpc_session.cc"
